requirements.txt
gunicorn
django-newsletter
sorl-thumbnail
django-debug-toolbar
django-anymail[mailgun]
django-extensions
